변경: DrawStrategy 추가
