+++ src/main/kotlin/blackjack/domain/deck/DrawStrategy.kt
@@ -1,0 +1,7 @@
+package blackjack.domain.deck
+
+import blackjack.domain.card.Card
+
+fun interface DrawStrategy {
+    fun draw(cards: MutableSet<Card>): Card
+}
