+++ src/test/kotlin/blackjack/domain/game/StringParserTest.kt
@@ -1,6 +1,6 @@
 package blackjack.domain.game
 
-import blackjack.domain.card.Deck
+import blackjack.domain.deck.Deck
 import io.kotest.core.spec.style.StringSpec
 import io.kotest.matchers.shouldBe
